Back port of Bugfixing SSL tests

+++ build/certs.ps1
@@ -546,17 +546,21 @@ function install-root-ca-windows ( $filename, $add = $true ) {
 
         if ($add) {
             if ($existing.Count -eq 0) {
-            $store.Add($cert)
-            write-output "CERTS: added cert '$($cert.Subject)' to Cert:\LocalMachine\Root"
-        }
-        else {
+                $store.Add($cert)
+                write-output "CERTS: added cert '$($cert.Subject)' ($($cert.Thumbprint)) to Cert:\LocalMachine\Root"
+            }
+            else {
                 write-output "CERTS: found cert '$($cert.Subject)' ($($cert.Thumbprint)) already in Cert:\LocalMachine\Root"
             }
         }
         else {
             if ($existing.Count -eq 1) {
-            $store.Remove($cert)
-            write-output "CERTS: removed cert '$($cert.Subject)' from Cert:\LocalMachine\Root"
+                $store.Remove($cert)
+                write-output "CERTS: removed cert '$($cert.Subject)' ($($cert.Thumbprint)) from Cert:\LocalMachine\Root"
+            }
+            else {
+                write-output "CERTS: cert '$($cert.Subject)' ($($cert.Thumbprint)) not in Cert:\LocalMachine\Root"
+            }
         }
     }
     catch {
@@ -575,13 +579,24 @@ function install-root-ca-linux ( $filename, $add = $true ) {
     try {
         $fileonly = [IO.Path]::GetFileName($filename)
         if ($add) {
-            cp $filename /usr/local/share/ca-certificates/
-            update-ca-certificates
+            if (-not (test-path "/usr/local/share/ca-certificates/$fileonly")) {
+                cp $filename /usr/local/share/ca-certificates/
+                update-ca-certificates
+                write-output "CERTS: added cert '$fileonly' to /usr/local/share/ca-certificates"
+            }
+            else {
+                write-output "CERTS: found cert '$fileonly' already in /usr/local/share/ca-certificates"
+            }
         }
         else {
-            $fileonly = [IO.Path]::GetFileName($filename)
-            rm /usr/local/share/ca-certificates/$fileonly
-            update-ca-certificates --fresh
+            if (test-path "/usr/local/share/ca-certificates/$fileonly") {
+                rm /usr/local/share/ca-certificates/$fileonly
+                update-ca-certificates --fresh
+                write-output "CERTS: removed cert '$fileonly' from /usr/local/share/ca-certificates"
+            }
+            else {
+                write-output "CERTS: cert '$fileonly' not in /usr/local/share/ca-certificates"
+            }
         }
         $global:CERTSEXITCODE = 0
     }
@@ -605,7 +620,7 @@ function install-root-ca ( $filename, $add = $true ) {
     }
 }
 
-function remove-root-ca ( $fiename ) {
+function remove-root-ca ( $filename ) {
     install-root-ca $filename $false
 }
